chore(nx-dev): rename aws file
