+++ nx-dev/ui-icons/src/index.ts
@@ -11,7 +11,7 @@ export * from './lib/customers/7eleven';
 export * from './lib/customers/adidas';
 export * from './lib/customers/adobe';
 export * from './lib/customers/american-airlines';
-export * from './lib/customers/aws-amplify';
+export * from './lib/customers/aws';
 export * from './lib/customers/bill';
 export * from './lib/customers/bloomberg';
 export * from './lib/customers/builder-mitosis';
