+++ nx-dev/ui-icons/src/lib/customers/aws.tsx
